wlan: Change the behavior of isLoadUnloadInProgres
Some APIs (wlan_hdd_linux_reg_notifier) can be allowed to execute if
driver load is in progress but should be blocked if driver unload is in
progress. To be able to cater to the needs of such APIs change
isLoadUnloadInProgress so that it can distinguish between driver load
and unload.

Change-Id: I80d42a2b1cc2e161c5405bcc8440f3fb02c6d4a7
CRs-Fixed: 629537