wlan: do not read registers if SSR is in progress
If WCNSS has crashed or is in bad state we should not
read registers in the WCNSS domain as it would lead
to watchdog bites because wpalReadRegister() does not
return till the register is read successfully. Add a
check to make sure that the driver does not try to
access registers when WCNSS is in a bad state.

Change-Id: Iafeb0ba8efcc5d6ec1c53939a6fc8ce5252553a4
CRs-Fixed: 621974