wlan: prima: properly handle SETBAND wext ioctl
The wlan driver has multiple mechanisms for configuring the current
operating band including the SETBAND driver command and the SETBAND
wext ioctl. Currently both the driver command and the wext ioctl call
the same hdd_setBand_helper() API which expects an ASCII command.
This is appropriate for the driver command since it uses an ASCII
string, but this is not appropriate for the wext ioctl since the data
is passed as an integer.  Create a new hdd_setBand() API which takes
binary data, and modify both hdd_setBand_helper() and the SETBAND wext
ioctl to invoke this API.  Also clean up other places where
hdd_setBand_helper() is being called with an ASCII command when it is
more efficient to call hdd_setBand() with an integer parameter.
set band helper api needs "SETBAND" as string to work properly but
from iwpriv command only the band number is received. band number
should be appended to "SETBAND" string manually.

Change-Id: I900618bea77c0eb46235eb11d33ad8aa22bde5a2
CRs-fixed: 642143