wlan: Fix SA Query Timeout
When in station mode, with a PMF protected connection to the AP,
if an SA Query procedure is initiated with the AP then the procedure
always times out. This is true even if the PMF connection with the
AP is still active and both the station and the AP still have the
correct security context.

This is fixed by sending the SA Query Response to the supplicant
when the device is in station mode instead of having the driver
handle the SA Query Response.

Change-Id: I1413e5749991e173ae9642b09f6771c158cc4bda
CRs-Fixed: 649809