wlan: Correct the Fix for SA Query Timeout
A fix was made for a problem that occurred when in station mode
where an SA Query procedure initiated with the AP would always
time out.  That fix was incorrect.  It fixed the problem when in
station mode, but broke the SA Query procedure when in soft AP
mode.

The problem was that all SA Query Response frames were
registered to be passed up to the supplicant or hostapd.  In fact,
SA Query Response frames should only be delivered to the
supplicant when in station mode.  This fix insures that SA Query
Response frames are passed up only in station mode and allows
them to be processed in LIM for soft AP mode.

Change-Id: I018a306e2aff1968cd0d5dc59ab58e3cd60ece8b
CRs-Fixed: 649809