wlan: Cannot change PMF SA Query configuration
The PMF SA Query configuration parameters "maximum
retries" and "retry interval" cannot be changed by
setting to new values in the WCNSS_qcom_cfg.ini
file.

The problem is in the definition of these parameters
in the configuration database.  They were defined as
read-only, when they should have been defined as
read-write.

Because of this issue, the driver always used the
default values for the parameters instead of the
new values.

Change-Id: I6e7dad199f619626fc2d06b2a6e1669ececd1c07
CRs-Fixed: 665295, 669811