wlan: From HDD remove the logic to try shared auth if open fails.
HDD tries to connect using shared auth if connect with open auth
fails. This leads driver to send 2 Assoc reject to the
supplicant which may lead to mismatch in states of driver and
supplicant.

In LIM if auth type is eCSR_AUTH_TYPE_AUTOSWITCH and if the open
auth is rejected with reason code
eSIR_MAC_AUTH_ALGO_NOT_SUPPORTED_STATUS (13) it try again with
shared auth.

Removes the logic from HDD as it is already handled by LIM.

Change-Id: I502c21448cfbb6f9fbfe6c8b918409c0cc96232d
CRs-Fixed: 679773